case-manager/pyproject.toml
[project]
name = "case-manager"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-xray-sdk>=2.15.0",
    "boto3>=1.42.89",
    "cachetools>=5.5.0",
    "django>=5.2.2",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-iam-dbauth>=0.2.1",
    "django-simple-history>=3.8.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-camel-case>=1.4.2",
    "drf-spectacular>=0.28.0",
    "psycopg[binary]>=3.2.9",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "regex>=2024.11.6",
    "requests>=2.32.5",
    "serverless-wsgi>=3.0.5",
    "six>=1.17.0",
    "ulid-py>=1.1.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
     { include-group = "test" },
    "black>=26.3.1",
    "check-jsonschema>=0.37.1",
    "datamodel-code-generator>=0.54.1",
    "django-extensions>=4.1",
]
test = [
    "boto3>=1.42.89",
    "coverage>=7.8.2",
    "factory-boy>=3.3.3",
    "mockito>=1.5.4",
    "pytest>=8.4.0",
    "pytz>=2025.2",
]
